+++ Classes/MovingEntity.hpp
@@ -24,10 +24,10 @@ namespace realtrick
         cocos2d::Vec2               _heading;
         cocos2d::Vec2               _side;
     
-        double                      _mass;
-        double                      _maxSpeed;
-        double                      _maxForce;
-        double                      _maxTurnRate;
+        float                       _mass;
+        float                       _maxSpeed;
+        float                       _maxForce;
+        float                       _maxTurnRate;
     
     public:
     
@@ -38,24 +38,24 @@ namespace realtrick
     
         cocos2d::Vec2 getVelocity() const                           { return _velocity; }
         void setVelocity(const cocos2d::Vec2& velocity)             { _velocity = velocity; }
-        double getSpeed() const                                     { return _velocity.getLength(); }
+        float getSpeed() const                                      { return _velocity.getLength(); }
         
         cocos2d::Vec2 getHeading() const                            { return _heading; }
         inline void setHeading(const cocos2d::Vec2& heading);
         
         cocos2d::Vec2 getSide() const                               { return _side; }
         
-        double getMass() const                                      { return _mass; }
-        void setMass(double mass)                                   { _mass = mass; }
+        float getMass() const                                       { return _mass; }
+        void setMass(float mass)                                    { _mass = mass; }
         
-        double getMaxSpeed() const                                  { return _maxSpeed; }
-        void setMaxSpeed(double maxSpeed)                           { _maxSpeed = maxSpeed; }
+        float getMaxSpeed() const                                   { return _maxSpeed; }
+        void setMaxSpeed(float maxSpeed)                            { _maxSpeed = maxSpeed; }
         
-        double getMaxForce() const                                  { return _maxForce; }
-        void setMaxForce(double maxForce)                           { _maxForce = maxForce; }
+        float getMaxForce() const                                   { return _maxForce; }
+        void setMaxForce(float maxForce)                            { _maxForce = maxForce; }
         
-        double getMaxTurnRate() const                               { return _maxTurnRate; }
-        void setMaxTurnRate(double maxTurnRate)                     { _maxTurnRate = maxTurnRate; }
+        float getMaxTurnRate() const                                { return _maxTurnRate; }
+        void setMaxTurnRate(float maxTurnRate)                      { _maxTurnRate = maxTurnRate; }
         
         virtual bool handleMessage(const Telegram& msg) override    { return false; }
         virtual void update(float dt) override {}
